+++ angel_orm/lib/src/annotations.dart
@@ -2,12 +2,12 @@ const Orm orm = const Orm();
 
 class Orm {
   /// The name of the table to query.
-  /// 
+  ///
   /// Inferred if not present.
   final String tableName;
-  
+
   /// Whether to generate migrations for this model.
-  /// 
+  ///
   /// Defaults to [:true:].
   final bool generateMigrations;